requirements.txt
pynubank==2.21.1
python-dotenv==1.0.0
alembic==1.10.2
psycopg2==2.9.5
python-dateutil==2.8.2
pytz==2022.7.1